Logramos compilar y ejecutar el programa de manera exitosa, generamos todos los archivos necesarios de las posiciones para posteriormente generar las gráficas correspondientes. Saludos

+++ distancias.c
@@ -3,13 +3,13 @@
 #include<math.h>
 #include<stdlib.h>
 
-double distancia(double x1,double x2)
+long double distancia(long double x1,long double x2)
 {
 
 return x2-x1;
 }
 
-double distanciat(double dx,double dy,double dz)
+long double distanciat(long double dx,long double dy,long double dz)
 {
 
 return sqrt(pow(dx,2)+pow(dy,2)+pow(dz,2));
