+++ posicionesfunc.c
@@ -4,7 +4,7 @@
 #include<stdlib.h>
 
 
-double posicion(double xp,double h,double vxp,double fxp)
+long double posicion(long double xp,long double h,long double vxp,long double fxp)
 {
 	return xp+(h*vxp)+((1/2)*fxp*(pow(h,2)));
 }
